+++ components/raftstore/src/store/peer_storage.rs
@@ -2,23 +2,18 @@
 
 // #[PerformanceCriticalPath]
 use std::{
-    cell::{Cell, RefCell},
-    cmp,
-    collections::VecDeque,
-    error, mem,
-    ops::{Deref, DerefMut, Range},
+    cell::RefCell,
+    error,
+    ops::{Deref, DerefMut},
     sync::{
         atomic::{AtomicBool, AtomicU64, AtomicUsize, Ordering},
         mpsc::{self, Receiver, TryRecvError},
-        Arc, Mutex,
+        Arc,
     },
     u64,
 };
 
-use collections::HashMap;
-use engine_traits::{
-    Engines, KvEngine, Mutable, Peekable, RaftEngine, RaftLogBatch, CF_RAFT, RAFT_LOG_MULTI_GET_CNT,
-};
+use engine_traits::{Engines, KvEngine, Mutable, Peekable, RaftEngine, RaftLogBatch, CF_RAFT};
 use fail::fail_point;
 use into_other::into_other;
 use keys::{self, enc_end_key, enc_start_key};
@@ -32,20 +27,20 @@ use protobuf::Message;
 use raft::{
     self,
     eraftpb::{self, ConfState, Entry, HardState, Snapshot},
-    util::limit_size,
     Error as RaftError, GetEntriesContext, RaftState, Ready, Storage, StorageError,
 };
-use tikv_alloc::trace::TraceEvent;
 use tikv_util::{
     box_err, box_try, debug, defer, error, info, time::Instant, warn, worker::Scheduler,
 };
 
-use super::{metrics::*, worker::RegionTask, SnapEntry, SnapKey, SnapManager, SnapshotStatistics};
+use super::{
+    entry_storage::last_index, metrics::*, worker::RegionTask, SnapEntry, SnapKey, SnapManager,
+    SnapshotStatistics,
+};
 use crate::{
-    bytes_capacity,
     store::{
-        async_io::write::WriteTask, fsm::GenSnapTask, memory::*, peer::PersistSnapshotResult, util,
-        worker::RaftlogFetchTask,
+        async_io::write::WriteTask, entry_storage::EntryStorage, fsm::GenSnapTask,
+        peer::PersistSnapshotResult, util, worker::RaftlogFetchTask,
     },
     Error, Result,
 };
@@ -55,16 +50,11 @@ use crate::{
 pub const RAFT_INIT_LOG_TERM: u64 = 5;
 pub const RAFT_INIT_LOG_INDEX: u64 = 5;
 const MAX_SNAP_TRY_CNT: usize = 5;
-const MAX_ASYNC_FETCH_TRY_CNT: usize = 3;
-
-pub const MAX_INIT_ENTRY_COUNT: usize = 1024;
 
 /// The initial region epoch version.
 pub const INIT_EPOCH_VER: u64 = 1;
 /// The initial region epoch conf_version.
 pub const INIT_EPOCH_CONF_VER: u64 = 1;
-
-const SHRINK_CACHE_CAPACITY: usize = 64;
 
 pub const JOB_STATUS_PENDING: usize = 0;
 pub const JOB_STATUS_RUNNING: usize = 1;
@@ -72,8 +62,6 @@ pub const JOB_STATUS_CANCELLING: usize = 2;
 pub const JOB_STATUS_CANCELLED: usize = 3;
 pub const JOB_STATUS_FINISHED: usize = 4;
 pub const JOB_STATUS_FAILED: usize = 5;
-
-const ENTRY_MEM_SIZE: usize = mem::size_of::<Entry>();
 
 /// Possible status returned by `check_applying_snap`.
 #[derive(Debug, Clone, Copy, PartialEq)]
@@ -112,749 +100,7 @@ impl PartialEq for SnapState {
     }
 }
 
-#[inline]
-pub fn first_index(state: &RaftApplyState) -> u64 {
-    state.get_truncated_state().get_index() + 1
-}
-
-#[inline]
-pub fn last_index(state: &RaftLocalState) -> u64 {
-    state.get_last_index()
-}
-
-struct EntryCache {
-    // The last index of persisted entry.
-    // It should be equal to `RaftLog::persisted`.
-    persisted: u64,
-    cache: VecDeque<Entry>,
-    trace: VecDeque<CachedEntries>,
-    hit: Cell<u64>,
-    miss: Cell<u64>,
-    #[cfg(test)]
-    size_change_cb: Option<Box<dyn Fn(i64) + Send + 'static>>,
-}
-
-impl EntryCache {
-    fn first_index(&self) -> Option<u64> {
-        self.cache.front().map(|e| e.get_index())
-    }
-
-    fn fetch_entries_to(
-        &self,
-        begin: u64,
-        end: u64,
-        mut fetched_size: u64,
-        max_size: u64,
-        ents: &mut Vec<Entry>,
-    ) {
-        if begin >= end {
-            return;
-        }
-        assert!(!self.cache.is_empty());
-        let cache_low = self.cache.front().unwrap().get_index();
-        let start_idx = begin.checked_sub(cache_low).unwrap() as usize;
-        let limit_idx = end.checked_sub(cache_low).unwrap() as usize;
-
-        let mut end_idx = start_idx;
-        self.cache
-            .iter()
-            .skip(start_idx)
-            .take_while(|e| {
-                let cur_idx = end_idx as u64 + cache_low;
-                assert_eq!(e.get_index(), cur_idx);
-                let m = u64::from(e.compute_size());
-                fetched_size += m;
-                if fetched_size == m {
-                    end_idx += 1;
-                    fetched_size <= max_size && end_idx < limit_idx
-                } else if fetched_size <= max_size {
-                    end_idx += 1;
-                    end_idx < limit_idx
-                } else {
-                    false
-                }
-            })
-            .count();
-        // Cache either is empty or contains latest log. Hence we don't need to fetch log
-        // from rocksdb anymore.
-        assert!(end_idx == limit_idx || fetched_size > max_size);
-        let (first, second) = tikv_util::slices_in_range(&self.cache, start_idx, end_idx);
-        ents.extend_from_slice(first);
-        ents.extend_from_slice(second);
-    }
-
-    fn append(&mut self, region_id: u64, peer_id: u64, entries: &[Entry]) {
-        if !entries.is_empty() {
-            let mut mem_size_change = 0;
-            let old_capacity = self.cache.capacity();
-            mem_size_change += self.append_impl(region_id, peer_id, entries);
-            let new_capacity = self.cache.capacity();
-            mem_size_change += Self::get_cache_vec_mem_size_change(new_capacity, old_capacity);
-            mem_size_change += self.shrink_if_necessary();
-            self.flush_mem_size_change(mem_size_change);
-        }
-    }
-
-    fn append_impl(&mut self, region_id: u64, peer_id: u64, entries: &[Entry]) -> i64 {
-        let mut mem_size_change = 0;
-
-        if let Some(cache_last_index) = self.cache.back().map(|e| e.get_index()) {
-            let first_index = entries[0].get_index();
-            if cache_last_index >= first_index {
-                let cache_len = self.cache.len();
-                let truncate_to = cache_len
-                    .checked_sub((cache_last_index - first_index + 1) as usize)
-                    .unwrap_or_default();
-                let trunc_to_idx = self.cache[truncate_to].index;
-                for e in self.cache.drain(truncate_to..) {
-                    mem_size_change -=
-                        (bytes_capacity(&e.data) + bytes_capacity(&e.context)) as i64;
-                }
-                if let Some(cached) = self.trace.back() {
-                    // Only committed entries can be traced, and only uncommitted entries
-                    // can be truncated. So there won't be any overlaps.
-                    let cached_last = cached.range.end - 1;
-                    assert!(cached_last < trunc_to_idx);
-                }
-            } else if cache_last_index + 1 < first_index {
-                panic!(
-                    "[region {}] {} unexpected hole: {} < {}",
-                    region_id, peer_id, cache_last_index, first_index
-                );
-            }
-        }
-
-        for e in entries {
-            self.cache.push_back(e.to_owned());
-            mem_size_change += (bytes_capacity(&e.data) + bytes_capacity(&e.context)) as i64;
-        }
-        // In the past, the entry cache will be truncated if its size exceeds a certain number.
-        // However, after introducing async write io, the entry must stay in cache if it's not
-        // persisted to raft db because the raft-rs may need to read entries.(e.g. leader sends
-        // MsgAppend to followers)
-
-        mem_size_change
-    }
-
-    pub fn entry(&self, idx: u64) -> Option<&Entry> {
-        let cache_low = self.cache.front()?.get_index();
-        if idx >= cache_low {
-            Some(&self.cache[(idx - cache_low) as usize])
-        } else {
-            None
-        }
-    }
-
-    /// Compact all entries whose indexes are less than `idx`.
-    pub fn compact_to(&mut self, mut idx: u64) -> u64 {
-        if idx > self.persisted + 1 {
-            // Only the persisted entries can be compacted
-            idx = self.persisted + 1;
-        }
-
-        let mut mem_size_change = 0;
-
-        // Clean cached entries which have been already sent to apply threads. For example,
-        // if entries [1, 10), [10, 20), [20, 30) are sent to apply threads and `compact_to(15)`
-        // is called, only [20, 30) will still be kept in cache.
-        let old_trace_cap = self.trace.capacity();
-        while let Some(cached_entries) = self.trace.pop_front() {
-            if cached_entries.range.start >= idx {
-                self.trace.push_front(cached_entries);
-                let trace_len = self.trace.len();
-                let trace_cap = self.trace.capacity();
-                if trace_len < SHRINK_CACHE_CAPACITY && trace_cap > SHRINK_CACHE_CAPACITY {
-                    self.trace.shrink_to(SHRINK_CACHE_CAPACITY);
-                }
-                break;
-            }
-            let (_, dangle_size) = cached_entries.take_entries();
-            mem_size_change -= dangle_size as i64;
-            idx = cmp::max(cached_entries.range.end, idx);
-        }
-        let new_trace_cap = self.trace.capacity();
-        mem_size_change += Self::get_trace_vec_mem_size_change(new_trace_cap, old_trace_cap);
-
-        let cache_first_idx = self.first_index().unwrap_or(u64::MAX);
-        if cache_first_idx >= idx {
-            self.flush_mem_size_change(mem_size_change);
-            assert!(mem_size_change <= 0);
-            return -mem_size_change as u64;
-        }
-
-        let cache_last_idx = self.cache.back().unwrap().get_index();
-        // Use `cache_last_idx + 1` to make sure cache can be cleared completely if necessary.
-        let compact_to = (cmp::min(cache_last_idx + 1, idx) - cache_first_idx) as usize;
-        for e in self.cache.drain(..compact_to) {
-            mem_size_change -= (bytes_capacity(&e.data) + bytes_capacity(&e.context)) as i64
-        }
-
-        mem_size_change += self.shrink_if_necessary();
-        self.flush_mem_size_change(mem_size_change);
-        assert!(mem_size_change <= 0);
-        -mem_size_change as u64
-    }
-
-    fn get_total_mem_size(&self) -> i64 {
-        let data_size: i64 = self
-            .cache
-            .iter()
-            .map(|e| (bytes_capacity(&e.data) + bytes_capacity(&e.context)) as i64)
-            .sum();
-        let cache_vec_size = Self::get_cache_vec_mem_size_change(self.cache.capacity(), 0);
-        let trace_vec_size = Self::get_trace_vec_mem_size_change(self.trace.capacity(), 0);
-        data_size + cache_vec_size + trace_vec_size
-    }
-
-    fn get_cache_vec_mem_size_change(new_capacity: usize, old_capacity: usize) -> i64 {
-        ENTRY_MEM_SIZE as i64 * (new_capacity as i64 - old_capacity as i64)
-    }
-
-    fn get_trace_vec_mem_size_change(new_capacity: usize, old_capacity: usize) -> i64 {
-        mem::size_of::<CachedEntries>() as i64 * (new_capacity as i64 - old_capacity as i64)
-    }
-
-    fn flush_mem_size_change(&self, mem_size_change: i64) {
-        #[cfg(test)]
-        if let Some(size_change_cb) = self.size_change_cb.as_ref() {
-            size_change_cb(mem_size_change);
-        }
-        let event = if mem_size_change > 0 {
-            TraceEvent::Add(mem_size_change as usize)
-        } else {
-            TraceEvent::Sub(-mem_size_change as usize)
-        };
-        MEMTRACE_ENTRY_CACHE.trace(event);
-        RAFT_ENTRIES_CACHES_GAUGE.add(mem_size_change);
-    }
-
-    fn flush_stats(&self) {
-        let hit = self.hit.replace(0);
-        RAFT_ENTRY_FETCHES.hit.inc_by(hit);
-        let miss = self.miss.replace(0);
-        RAFT_ENTRY_FETCHES.miss.inc_by(miss);
-    }
-
-    #[inline]
-    fn is_empty(&self) -> bool {
-        self.cache.is_empty()
-    }
-
-    fn trace_cached_entries(&mut self, entries: CachedEntries) {
-        let dangle_size = {
-            let mut guard = entries.entries.lock().unwrap();
-
-            let last_idx = guard.0.last().map(|e| e.index).unwrap();
-            let cache_front = match self.cache.front().map(|e| e.index) {
-                Some(i) => i,
-                None => u64::MAX,
-            };
-
-            let dangle_range = if last_idx < cache_front {
-                // All entries are not in entry cache.
-                0..guard.0.len()
-            } else if let Ok(i) = guard.0.binary_search_by(|e| e.index.cmp(&cache_front)) {
-                // Some entries are in entry cache.
-                0..i
-            } else {
-                // All entries are in entry cache.
-                0..0
-            };
-
-            let mut size = 0;
-            for e in &guard.0[dangle_range] {
-                size += bytes_capacity(&e.data) + bytes_capacity(&e.context);
-            }
-            guard.1 = size;
-            size
-        };
-
-        let old_capacity = self.trace.capacity();
-        self.trace.push_back(entries);
-        let new_capacity = self.trace.capacity();
-        let diff = Self::get_trace_vec_mem_size_change(new_capacity, old_capacity);
-
-        self.flush_mem_size_change(diff + dangle_size as i64);
-    }
-
-    fn shrink_if_necessary(&mut self) -> i64 {
-        if self.cache.len() < SHRINK_CACHE_CAPACITY && self.cache.capacity() > SHRINK_CACHE_CAPACITY
-        {
-            let old_capacity = self.cache.capacity();
-            self.cache.shrink_to_fit();
-            let new_capacity = self.cache.capacity();
-            return Self::get_cache_vec_mem_size_change(new_capacity, old_capacity);
-        }
-        0
-    }
-
-    fn update_persisted(&mut self, persisted: u64) {
-        self.persisted = persisted;
-    }
-}
-
-impl Default for EntryCache {
-    fn default() -> Self {
-        let entry_cache = EntryCache {
-            persisted: 0,
-            cache: Default::default(),
-            trace: Default::default(),
-            hit: Cell::new(0),
-            miss: Cell::new(0),
-            #[cfg(test)]
-            size_change_cb: None,
-        };
-        entry_cache.flush_mem_size_change(entry_cache.get_total_mem_size());
-        entry_cache
-    }
-}
-
-impl Drop for EntryCache {
-    fn drop(&mut self) {
-        let mem_size_change = self.get_total_mem_size();
-        self.flush_mem_size_change(-mem_size_change);
-        self.flush_stats();
-    }
-}
-
-/// A subset of `PeerStorage` that focus on accessing log entries.
-pub struct EntryStorage<ER> {
-    region_id: u64,
-    peer_id: u64,
-    raft_engine: ER,
-    cache: EntryCache,
-    raft_state: RaftLocalState,
-    apply_state: RaftApplyState,
-    last_term: u64,
-    raftlog_fetch_scheduler: Scheduler<RaftlogFetchTask>,
-    raftlog_fetch_stats: AsyncFetchStats,
-    async_fetch_results: RefCell<HashMap<u64, RaftlogFetchState>>,
-}
-
-impl<ER: RaftEngine> EntryStorage<ER> {
-    pub fn new(
-        region_id: u64,
-        peer_id: u64,
-        raft_engine: ER,
-        raft_state: RaftLocalState,
-        apply_state: RaftApplyState,
-        last_term: u64,
-        raftlog_fetch_scheduler: Scheduler<RaftlogFetchTask>,
-    ) -> Self {
-        EntryStorage {
-            region_id,
-            peer_id,
-            raft_engine,
-            cache: EntryCache::default(),
-            raft_state,
-            apply_state,
-            last_term,
-            raftlog_fetch_scheduler,
-            raftlog_fetch_stats: AsyncFetchStats::default(),
-            async_fetch_results: RefCell::new(HashMap::default()),
-        }
-    }
-
-    fn check_range(&self, low: u64, high: u64) -> raft::Result<()> {
-        if low > high {
-            return Err(storage_error(format!(
-                "low: {} is greater that high: {}",
-                low, high
-            )));
-        } else if low <= self.truncated_index() {
-            return Err(RaftError::Store(StorageError::Compacted));
-        } else if high > self.last_index() + 1 {
-            return Err(storage_error(format!(
-                "entries' high {} is out of bound lastindex {}",
-                high,
-                self.last_index()
-            )));
-        }
-        Ok(())
-    }
-
-    pub fn clean_async_fetch_res(&mut self, low: u64) {
-        self.async_fetch_results.borrow_mut().remove(&low);
-    }
-
-    // Update the async fetch result.
-    // None indicates cleanning the fetched result.
-    pub fn update_async_fetch_res(&mut self, low: u64, res: Option<Box<RaftlogFetchResult>>) {
-        // If it's in fetching, don't clean the async fetch result.
-        if self.async_fetch_results.borrow().get(&low) == Some(&RaftlogFetchState::Fetching)
-            && res.is_none()
-        {
-            return;
-        }
-
-        match res {
-            Some(res) => {
-                if let Some(RaftlogFetchState::Fetched(prev)) = self
-                    .async_fetch_results
-                    .borrow_mut()
-                    .insert(low, RaftlogFetchState::Fetched(res))
-                {
-                    info!(
-                        "unconsumed async fetch res";
-                        "region_id" => self.region_id,
-                        "peer_id" => self.peer_id,
-                        "res" => ?prev,
-                        "low" => low,
-                    );
-                }
-            }
-            None => {
-                let prev = self.async_fetch_results.borrow_mut().remove(&low);
-                if prev.is_some() {
-                    self.raftlog_fetch_stats.fetch_unused.update(|m| m + 1);
-                }
-            }
-        }
-    }
-
-    fn async_fetch(
-        &self,
-        region_id: u64,
-        low: u64,
-        high: u64,
-        max_size: u64,
-        context: GetEntriesContext,
-        buf: &mut Vec<Entry>,
-    ) -> raft::Result<usize> {
-        if let Some(RaftlogFetchState::Fetching) = self.async_fetch_results.borrow().get(&low) {
-            // already an async fetch in flight
-            return Err(raft::Error::Store(
-                raft::StorageError::LogTemporarilyUnavailable,
-            ));
-        }
-
-        let tried_cnt = if let Some(RaftlogFetchState::Fetched(res)) =
-            self.async_fetch_results.borrow_mut().remove(&low)
-        {
-            assert_eq!(res.low, low);
-            let mut ents = res.ents?;
-            let first = ents.first().map(|e| e.index).unwrap();
-            assert_eq!(first, res.low);
-            let last = ents.last().map(|e| e.index).unwrap();
-
-            if last + 1 >= high {
-                // async fetch res covers [low, high)
-                ents.truncate((high - first) as usize);
-                assert_eq!(ents.last().map(|e| e.index).unwrap(), high - 1);
-                if max_size < res.max_size {
-                    limit_size(&mut ents, Some(max_size));
-                }
-                let count = ents.len();
-                buf.append(&mut ents);
-                fail_point!("on_async_fetch_return");
-                return Ok(count);
-            } else if res.hit_size_limit && max_size <= res.max_size {
-                // async fetch res doesn't cover [low, high) due to hit size limit
-                if max_size < res.max_size {
-                    limit_size(&mut ents, Some(max_size));
-                };
-                let count = ents.len();
-                buf.append(&mut ents);
-                return Ok(count);
-            } else if last + RAFT_LOG_MULTI_GET_CNT > high - 1
-                && res.tried_cnt + 1 == MAX_ASYNC_FETCH_TRY_CNT
-            {
-                let mut fetched_size = ents.iter().fold(0, |acc, e| acc + e.compute_size() as u64);
-                if max_size <= fetched_size {
-                    limit_size(&mut ents, Some(max_size));
-                    let count = ents.len();
-                    buf.append(&mut ents);
-                    return Ok(count);
-                }
-
-                // the count of left entries isn't too large, fetch the remaining entries synchronously one by one
-                for idx in last + 1..high {
-                    let ent = self.raft_engine.get_entry(region_id, idx)?;
-                    match ent {
-                        None => {
-                            return Err(raft::Error::Store(raft::StorageError::Unavailable));
-                        }
-                        Some(ent) => {
-                            let size = ent.compute_size() as u64;
-                            if fetched_size + size > max_size {
-                                break;
-                            } else {
-                                fetched_size += size;
-                                ents.push(ent);
-                            }
-                        }
-                    }
-                }
-                let count = ents.len();
-                buf.append(&mut ents);
-                return Ok(count);
-            }
-            info!(
-                "async fetch invalid";
-                "region_id" => self.region_id,
-                "peer_id" => self.peer_id,
-                "first" => first,
-                "last" => last,
-                "low" => low,
-                "high" => high,
-                "max_size" => max_size,
-                "res_max_size" => res.max_size,
-            );
-            // low index or max size is changed, the result is not fit for the current range, so refetch again.
-            self.raftlog_fetch_stats.fetch_invalid.update(|m| m + 1);
-            res.tried_cnt + 1
-        } else {
-            1
-        };
-
-        // the first/second try: get [low, high) asynchronously
-        // the third try:
-        //  - if term and low are matched: use result of [low, persisted) and get [persisted, high) synchronously
-        //  - else: get [low, high) synchronously
-        if tried_cnt >= MAX_ASYNC_FETCH_TRY_CNT {
-            // even the larger range is invalid again, fallback to fetch in sync way
-            self.raftlog_fetch_stats.fallback_fetch.update(|m| m + 1);
-            let count = self.raft_engine.fetch_entries_to(
-                region_id,
-                low,
-                high,
-                Some(max_size as usize),
-                buf,
-            )?;
-            return Ok(count);
-        }
-
-        self.raftlog_fetch_stats.async_fetch.update(|m| m + 1);
-        self.async_fetch_results
-            .borrow_mut()
-            .insert(low, RaftlogFetchState::Fetching);
-        self.raftlog_fetch_scheduler
-            .schedule(RaftlogFetchTask::PeerStorage {
-                region_id,
-                context,
-                low,
-                high,
-                max_size: (max_size as usize),
-                tried_cnt,
-                term: self.hard_state().get_term(),
-            })
-            .unwrap();
-        Err(raft::Error::Store(
-            raft::StorageError::LogTemporarilyUnavailable,
-        ))
-    }
-
-    pub fn entries(
-        &self,
-        low: u64,
-        high: u64,
-        max_size: u64,
-        context: GetEntriesContext,
-    ) -> raft::Result<Vec<Entry>> {
-        self.check_range(low, high)?;
-        let mut ents =
-            Vec::with_capacity(std::cmp::min((high - low) as usize, MAX_INIT_ENTRY_COUNT));
-        if low == high {
-            return Ok(ents);
-        }
-        let cache_low = self.cache.first_index().unwrap_or(u64::MAX);
-        if high <= cache_low {
-            self.cache.miss.update(|m| m + 1);
-            return if context.can_async() {
-                self.async_fetch(self.region_id, low, high, max_size, context, &mut ents)?;
-                Ok(ents)
-            } else {
-                self.raftlog_fetch_stats.sync_fetch.update(|m| m + 1);
-                self.raft_engine.fetch_entries_to(
-                    self.region_id,
-                    low,
-                    high,
-                    Some(max_size as usize),
-                    &mut ents,
-                )?;
-                Ok(ents)
-            };
-        }
-        let begin_idx = if low < cache_low {
-            self.cache.miss.update(|m| m + 1);
-            let fetched_count = if context.can_async() {
-                self.async_fetch(self.region_id, low, cache_low, max_size, context, &mut ents)?
-            } else {
-                self.raftlog_fetch_stats.sync_fetch.update(|m| m + 1);
-                self.raft_engine.fetch_entries_to(
-                    self.region_id,
-                    low,
-                    cache_low,
-                    Some(max_size as usize),
-                    &mut ents,
-                )?
-            };
-            if fetched_count < (cache_low - low) as usize {
-                // Less entries are fetched than expected.
-                return Ok(ents);
-            }
-            cache_low
-        } else {
-            low
-        };
-        self.cache.hit.update(|h| h + 1);
-        let fetched_size = ents.iter().fold(0, |acc, e| acc + e.compute_size());
-        self.cache
-            .fetch_entries_to(begin_idx, high, fetched_size as u64, max_size, &mut ents);
-        Ok(ents)
-    }
-
-    pub fn term(&self, idx: u64) -> raft::Result<u64> {
-        if idx == self.truncated_index() {
-            return Ok(self.truncated_term());
-        }
-        self.check_range(idx, idx + 1)?;
-        if self.truncated_term() == self.last_term || idx == self.last_index() {
-            return Ok(self.last_term);
-        }
-        if let Some(e) = self.cache.entry(idx) {
-            Ok(e.get_term())
-        } else {
-            Ok(self
-                .raft_engine
-                .get_entry(self.region_id, idx)
-                .unwrap()
-                .unwrap()
-                .get_term())
-        }
-    }
-
-    #[inline]
-    pub fn first_index(&self) -> u64 {
-        first_index(&self.apply_state)
-    }
-
-    #[inline]
-    pub fn last_index(&self) -> u64 {
-        last_index(&self.raft_state)
-    }
-
-    #[inline]
-    pub fn last_term(&self) -> u64 {
-        self.last_term
-    }
-
-    #[inline]
-    pub fn raft_state(&self) -> &RaftLocalState {
-        &self.raft_state
-    }
-
-    #[inline]
-    pub fn applied_index(&self) -> u64 {
-        self.apply_state.get_applied_index()
-    }
-
-    #[inline]
-    pub fn set_applied_state(&mut self, apply_state: RaftApplyState) {
-        self.apply_state = apply_state;
-    }
-
-    #[inline]
-    pub fn apply_state(&self) -> &RaftApplyState {
-        &self.apply_state
-    }
-
-    #[inline]
-    pub fn commit_index(&self) -> u64 {
-        self.raft_state.get_hard_state().get_commit()
-    }
-
-    #[inline]
-    pub fn set_commit_index(&mut self, commit: u64) {
-        assert!(commit >= self.commit_index());
-        self.raft_state.mut_hard_state().set_commit(commit);
-    }
-
-    #[inline]
-    pub fn hard_state(&self) -> &HardState {
-        self.raft_state.get_hard_state()
-    }
-
-    #[inline]
-    pub fn truncated_index(&self) -> u64 {
-        self.apply_state.get_truncated_state().get_index()
-    }
-
-    #[inline]
-    pub fn truncated_term(&self) -> u64 {
-        self.apply_state.get_truncated_state().get_term()
-    }
-
-    // Append the given entries to the raft log using previous last index or self.last_index.
-    pub fn append<EK: KvEngine>(&mut self, entries: Vec<Entry>, task: &mut WriteTask<EK, ER>) {
-        if entries.is_empty() {
-            return;
-        }
-        debug!(
-            "append entries";
-            "region_id" => self.region_id,
-            "peer_id" => self.peer_id,
-            "count" => entries.len(),
-        );
-        let prev_last_index = self.raft_state.get_last_index();
-
-        let (last_index, last_term) = {
-            let e = entries.last().unwrap();
-            (e.get_index(), e.get_term())
-        };
-
-        self.cache.append(self.region_id, self.peer_id, &entries);
-
-        task.entries = entries;
-        // Delete any previously appended log entries which never committed.
-        task.cut_logs = Some((last_index + 1, prev_last_index + 1));
-
-        self.raft_state.set_last_index(last_index);
-        self.last_term = last_term;
-    }
-
-    pub fn compact_entry_cache(&mut self, idx: u64) {
-        self.cache.compact_to(idx);
-    }
-
-    #[inline]
-    pub fn is_entry_cache_empty(&self) -> bool {
-        self.cache.is_empty()
-    }
-
-    /// Evict entries from the cache.
-    pub fn evict_entry_cache(&mut self, half: bool) {
-        if !self.is_entry_cache_empty() {
-            let cache = &mut self.cache;
-            let cache_len = cache.cache.len();
-            let drain_to = if half { cache_len / 2 } else { cache_len - 1 };
-            let idx = cache.cache[drain_to].index;
-            let mem_size_change = cache.compact_to(idx + 1);
-            RAFT_ENTRIES_EVICT_BYTES.inc_by(mem_size_change);
-        }
-    }
-
-    #[inline]
-    pub fn flush_entry_cache_metrics(&mut self) {
-        // NOTE: memory usage of entry cache is flushed realtime.
-        self.cache.flush_stats();
-        self.raftlog_fetch_stats.flush_stats();
-    }
-
-    pub fn raft_engine(&self) -> &ER {
-        &self.raft_engine
-    }
-
-    pub fn update_cache_persisted(&mut self, persisted: u64) {
-        self.cache.update_persisted(persisted);
-    }
-
-    pub fn trace_cached_entries(&mut self, entries: CachedEntries) {
-        self.cache.trace_cached_entries(entries);
-    }
-}
-
-fn storage_error<E>(error: E) -> raft::Error
+pub fn storage_error<E>(error: E) -> raft::Error
 where
     E: Into<Box<dyn error::Error + Send + Sync>>,
 {
@@ -1091,56 +337,6 @@ where
     pub tag: String,
 }
 
-#[derive(Debug, PartialEq)]
-pub enum RaftlogFetchState {
-    Fetching,
-    Fetched(Box<RaftlogFetchResult>),
-}
-
-#[derive(Debug, PartialEq)]
-pub struct RaftlogFetchResult {
-    pub ents: raft::Result<Vec<Entry>>,
-    // because entries may be empty, so store the original low index that the task issued
-    pub low: u64,
-    // the original max size that the task issued
-    pub max_size: u64,
-    // if the ents hit max_size
-    pub hit_size_limit: bool,
-    // the times that async fetch have already tried
-    pub tried_cnt: usize,
-    // the term when the task issued
-    pub term: u64,
-}
-
-#[derive(Default)]
-struct AsyncFetchStats {
-    async_fetch: Cell<u64>,
-    sync_fetch: Cell<u64>,
-    fallback_fetch: Cell<u64>,
-    fetch_invalid: Cell<u64>,
-    fetch_unused: Cell<u64>,
-}
-
-impl AsyncFetchStats {
-    fn flush_stats(&mut self) {
-        RAFT_ENTRY_FETCHES
-            .async_fetch
-            .inc_by(self.async_fetch.replace(0));
-        RAFT_ENTRY_FETCHES
-            .sync_fetch
-            .inc_by(self.sync_fetch.replace(0));
-        RAFT_ENTRY_FETCHES
-            .fallback_fetch
-            .inc_by(self.fallback_fetch.replace(0));
-        RAFT_ENTRY_FETCHES
-            .fetch_invalid
-            .inc_by(self.fetch_invalid.replace(0));
-        RAFT_ENTRY_FETCHES
-            .fetch_unused
-            .inc_by(self.fetch_unused.replace(0));
-    }
-}
-
 impl<EK: KvEngine, ER: RaftEngine> Deref for PeerStorage<EK, ER> {
     type Target = EntryStorage<ER>;
 
@@ -1250,14 +446,14 @@ where
     }
 
     pub fn initial_state(&self) -> raft::Result<RaftState> {
-        let hard_state = self.raft_state.get_hard_state().clone();
+        let hard_state = self.raft_state().get_hard_state().clone();
         if hard_state == HardState::default() {
             assert!(
                 !self.is_initialized(),
                 "peer for region {:?} is initialized but local state {:?} has empty hard \
                  state",
                 self.region,
-                self.raft_state
+                self.raft_state()
             );
 
             return Ok(RaftState::new(hard_state, ConfState::default()));
@@ -1299,7 +495,7 @@ where
         snapshot_index: u64,
         kv_wb: &mut impl Mutable,
     ) -> Result<()> {
-        let mut snapshot_raft_state = self.raft_state.clone();
+        let mut snapshot_raft_state = self.raft_state().clone();
         snapshot_raft_state
             .mut_hard_state()
             .set_commit(snapshot_index);
@@ -1318,7 +514,7 @@ where
         kv_wb.put_msg_cf(
             CF_RAFT,
             &keys::apply_state_key(self.region.get_id()),
-            &self.apply_state,
+            self.apply_state(),
         )?;
         Ok(())
     }
@@ -1536,15 +732,17 @@ where
 
         let last_index = snap.get_metadata().get_index();
 
-        self.raft_state.set_last_index(last_index);
-        self.last_term = snap.get_metadata().get_term();
-        self.apply_state.set_applied_index(last_index);
-        self.applied_index_term = self.last_term;
+        self.raft_state_mut().set_last_index(last_index);
+        self.set_last_term(snap.get_metadata().get_term());
+        self.apply_state_mut().set_applied_index(last_index);
+        self.applied_index_term = self.last_term();
 
         // The snapshot only contains log which index > applied index, so
         // here the truncate state's (index, term) is in snapshot metadata.
-        self.apply_state.mut_truncated_state().set_index(last_index);
-        self.apply_state
+        self.apply_state_mut()
+            .mut_truncated_state()
+            .set_index(last_index);
+        self.apply_state_mut()
             .mut_truncated_state()
             .set_term(snap.get_metadata().get_term());
 
@@ -1563,7 +761,7 @@ where
             "region_id" => self.region.get_id(),
             "peer_id" => self.peer_id,
             "region" => ?region,
-            "state" => ?self.apply_state,
+            "state" => ?self.apply_state(),
         );
 
         Ok(region)
@@ -1583,9 +781,9 @@ where
             raft_wb,
             region_id,
             first_index,
-            &self.raft_state,
+            self.raft_state(),
         )?;
-        self.cache = EntryCache::default();
+        self.entry_storage.clear();
         Ok(())
     }
 
@@ -1782,7 +980,7 @@ where
         destroy_regions: Vec<metapb::Region>,
     ) -> Result<(HandleReadyResult, WriteTask<EK, ER>)> {
         let region_id = self.get_region_id();
-        let prev_raft_state = self.raft_state.clone();
+        let prev_raft_state = self.raft_state().clone();
 
         let mut write_task = WriteTask::new(region_id, self.peer_id, ready.number());
 
@@ -1808,15 +1006,15 @@ where
 
         // Last index is 0 means the peer is created from raft message
         // and has not applied snapshot yet, so skip persistent hard state.
-        if self.raft_state.get_last_index() > 0 {
+        if self.raft_state().get_last_index() > 0 {
             if let Some(hs) = ready.hs() {
-                self.raft_state.set_hard_state(hs.clone());
+                self.raft_state_mut().set_hard_state(hs.clone());
             }
         }
 
         // Save raft state if it has changed or there is a snapshot.
-        if prev_raft_state != self.raft_state || !ready.snapshot().is_empty() {
-            write_task.raft_state = Some(self.raft_state.clone());
+        if prev_raft_state != *self.raft_state() || !ready.snapshot().is_empty() {
+            write_task.raft_state = Some(self.raft_state().clone());
         }
 
         if !ready.snapshot().is_empty() {
@@ -2052,34 +1250,8 @@ pub fn write_peer_state<T: Mutable>(
     Ok(())
 }
 
-/// Committed entries sent to apply threads.
-#[derive(Clone)]
-pub struct CachedEntries {
-    pub range: Range<u64>,
-    // Entries and dangle size for them. `dangle` means not in entry cache.
-    entries: Arc<Mutex<(Vec<Entry>, usize)>>,
-}
-
-impl CachedEntries {
-    pub fn new(entries: Vec<Entry>) -> Self {
-        assert!(!entries.is_empty());
-        let start = entries.first().map(|x| x.index).unwrap();
-        let end = entries.last().map(|x| x.index).unwrap() + 1;
-        let range = Range { start, end };
-        CachedEntries {
-            entries: Arc::new(Mutex::new((entries, 0))),
-            range,
-        }
-    }
-
-    /// Take cached entries and dangle size for them. `dangle` means not in entry cache.
-    pub fn take_entries(&self) -> (Vec<Entry>, usize) {
-        mem::take(&mut *self.entries.lock().unwrap())
-    }
-}
-
 #[cfg(test)]
-mod tests {
+pub mod tests {
     use std::{
         cell::RefCell,
         path::Path,
@@ -2111,26 +1283,12 @@ mod tests {
         store::{
             async_io::write::write_to_db_for_test,
             bootstrap_store,
+            entry_storage::tests::validate_cache,
             fsm::apply::compact_raft_log,
             initial_region, prepare_bootstrap_cluster,
             worker::{RaftlogFetchRunner, RegionRunner, RegionTask},
         },
     };
-
-    impl EntryCache {
-        fn new_with_cb(cb: impl Fn(i64) + Send + 'static) -> Self {
-            let entry_cache = EntryCache {
-                persisted: 0,
-                cache: Default::default(),
-                trace: Default::default(),
-                hit: Cell::new(0),
-                miss: Cell::new(0),
-                size_change_cb: Some(Box::new(cb) as Box<dyn Fn(i64) + Send + 'static>),
-            };
-            entry_cache.flush_mem_size_change(entry_cache.get_total_mem_size());
-            entry_cache
-        }
-    }
 
     fn new_storage(
         region_scheduler: Scheduler<RegionTask<KvTestSnapshot>>,
@@ -2166,7 +1324,7 @@ mod tests {
         .unwrap()
     }
 
-    fn new_storage_from_ents(
+    pub fn new_storage_from_ents(
         region_scheduler: Scheduler<RegionTask<KvTestSnapshot>>,
         raftlog_fetch_scheduler: Scheduler<RaftlogFetchTask>,
         path: &TempDir,
@@ -2177,15 +1335,15 @@ mod tests {
         store.append(ents[1..].to_vec(), &mut write_task);
         store.update_cache_persisted(ents.last().unwrap().get_index());
         store
-            .apply_state
+            .apply_state_mut()
             .mut_truncated_state()
             .set_index(ents[0].get_index());
         store
-            .apply_state
+            .apply_state_mut()
             .mut_truncated_state()
             .set_term(ents[0].get_term());
         store
-            .apply_state
+            .apply_state_mut()
             .set_applied_index(ents.last().unwrap().get_index());
         if write_task.kv_wb.is_none() {
             write_task.kv_wb = Some(store.engines.kv.write_batch());
@@ -2193,35 +1351,22 @@ mod tests {
         store
             .save_apply_state_to(write_task.kv_wb.as_mut().unwrap())
             .unwrap();
-        write_task.raft_state = Some(store.raft_state.clone());
+        write_task.raft_state = Some(store.raft_state().clone());
         write_to_db_for_test(&store.engines, write_task);
         store
     }
 
-    fn append_ents(store: &mut PeerStorage<KvTestEngine, RaftTestEngine>, ents: &[Entry]) {
+    pub fn append_ents(store: &mut PeerStorage<KvTestEngine, RaftTestEngine>, ents: &[Entry]) {
         if ents.is_empty() {
             return;
         }
         let mut write_task = WriteTask::new(store.get_region_id(), store.peer_id, 1);
         store.append(ents.to_vec(), &mut write_task);
-        write_task.raft_state = Some(store.raft_state.clone());
+        write_task.raft_state = Some(store.raft_state().clone());
         write_to_db_for_test(&store.engines, write_task);
     }
 
-    fn validate_cache(store: &PeerStorage<KvTestEngine, RaftTestEngine>, exp_ents: &[Entry]) {
-        assert_eq!(store.cache.cache, exp_ents);
-        for e in exp_ents {
-            let entry = store
-                .engines
-                .raft
-                .get_entry(store.get_region_id(), e.get_index())
-                .unwrap()
-                .unwrap();
-            assert_eq!(entry, *e);
-        }
-    }
-
-    fn new_entry(index: u64, term: u64) -> Entry {
+    pub fn new_entry(index: u64, term: u64) -> Entry {
         let mut e = Entry::default();
         e.set_index(index);
         e.set_term(term);
@@ -2495,257 +1640,6 @@ mod tests {
         assert_ne!(count, 0);
     }
 
-    #[test]
-    fn test_async_fetch() {
-        let ents = vec![
-            new_entry(2, 2),
-            new_entry(3, 3),
-            new_entry(4, 4),
-            new_entry(5, 5),
-            new_entry(6, 6),
-        ];
-
-        let td = Builder::new().prefix("tikv-store-test").tempdir().unwrap();
-        let region_worker = Worker::new("snap-manager").lazy_build("snap-manager");
-        let region_scheduler = region_worker.scheduler();
-        let (dummy_scheduler, _rx) = dummy_scheduler();
-        let mut store = new_storage_from_ents(region_scheduler, dummy_scheduler, &td, &ents);
-
-        let max_u64 = u64::max_value();
-        let mut tests = vec![
-            // already compacted
-            (
-                3,
-                7,
-                max_u64,
-                1,
-                RaftlogFetchResult {
-                    ents: Err(RaftError::Store(StorageError::Compacted)),
-                    low: 3,
-                    max_size: max_u64,
-                    hit_size_limit: false,
-                    tried_cnt: 1,
-                    term: 1,
-                },
-                Err(RaftError::Store(StorageError::Compacted)),
-                vec![],
-            ),
-            // fetch partial entries due to max size limit
-            (
-                3,
-                7,
-                30,
-                1,
-                RaftlogFetchResult {
-                    ents: Ok(ents[1..4].to_vec()),
-                    low: 3,
-                    max_size: 30,
-                    hit_size_limit: true,
-                    tried_cnt: 1,
-                    term: 1,
-                },
-                Ok(3),
-                ents[1..4].to_vec(),
-            ),
-            // fetch all entries
-            (
-                2,
-                7,
-                max_u64,
-                1,
-                RaftlogFetchResult {
-                    ents: Ok(ents.clone()),
-                    low: 2,
-                    max_size: max_u64,
-                    hit_size_limit: false,
-                    tried_cnt: 1,
-                    term: 1,
-                },
-                Ok(5),
-                ents.clone(),
-            ),
-            // high is smaller than before
-            (
-                3,
-                5,
-                max_u64,
-                1,
-                RaftlogFetchResult {
-                    ents: Ok(ents[1..].to_vec()),
-                    low: 3,
-                    max_size: max_u64,
-                    hit_size_limit: false,
-                    tried_cnt: 1,
-                    term: 1,
-                },
-                Ok(2),
-                ents[1..3].to_vec(),
-            ),
-            // high is larger than before, second try
-            (
-                3,
-                7,
-                max_u64,
-                1,
-                RaftlogFetchResult {
-                    ents: Ok(ents[1..4].to_vec()),
-                    low: 3,
-                    max_size: max_u64,
-                    hit_size_limit: false,
-                    tried_cnt: 1,
-                    term: 1,
-                },
-                Err(RaftError::Store(StorageError::LogTemporarilyUnavailable)),
-                vec![],
-            ),
-            // high is larger than before, thrid try
-            (
-                3,
-                7,
-                max_u64,
-                1,
-                RaftlogFetchResult {
-                    ents: Ok(ents[1..4].to_vec()),
-                    low: 3,
-                    max_size: max_u64,
-                    hit_size_limit: false,
-                    tried_cnt: 2,
-                    term: 1,
-                },
-                Ok(4),
-                ents[1..].to_vec(),
-            ),
-            // max size is smaller than before
-            (
-                2,
-                7,
-                10,
-                1,
-                RaftlogFetchResult {
-                    ents: Ok(ents.clone()),
-                    low: 2,
-                    max_size: max_u64,
-                    hit_size_limit: false,
-                    tried_cnt: 1,
-                    term: 1,
-                },
-                Ok(2),
-                ents[..2].to_vec(),
-            ),
-            // max size is larger than before but with lower high
-            (
-                2,
-                5,
-                40,
-                1,
-                RaftlogFetchResult {
-                    ents: Ok(ents.clone()),
-                    low: 2,
-                    max_size: 30,
-                    hit_size_limit: false,
-                    tried_cnt: 1,
-                    term: 1,
-                },
-                Ok(3),
-                ents[..3].to_vec(),
-            ),
-            // low index is smaller than before
-            (
-                2,
-                7,
-                max_u64,
-                1,
-                RaftlogFetchResult {
-                    ents: Err(RaftError::Store(StorageError::Compacted)),
-                    low: 3,
-                    max_size: max_u64,
-                    hit_size_limit: false,
-                    tried_cnt: 1,
-                    term: 1,
-                },
-                Err(RaftError::Store(StorageError::LogTemporarilyUnavailable)),
-                vec![],
-            ),
-            // low index is larger than before
-            (
-                4,
-                7,
-                max_u64,
-                1,
-                RaftlogFetchResult {
-                    ents: Ok(vec![]),
-                    low: 3,
-                    max_size: max_u64,
-                    hit_size_limit: false,
-                    tried_cnt: 1,
-                    term: 1,
-                },
-                Err(RaftError::Store(StorageError::LogTemporarilyUnavailable)),
-                vec![],
-            ),
-            // hit tried several lmit
-            (
-                3,
-                7,
-                max_u64,
-                1,
-                RaftlogFetchResult {
-                    ents: Ok(ents[1..4].to_vec()),
-                    low: 3,
-                    max_size: max_u64,
-                    hit_size_limit: false,
-                    tried_cnt: MAX_ASYNC_FETCH_TRY_CNT,
-                    term: 1,
-                },
-                Ok(4),
-                ents[1..5].to_vec(),
-            ),
-            // term is changed
-            (
-                3,
-                7,
-                max_u64,
-                2,
-                RaftlogFetchResult {
-                    ents: Ok(ents[1..4].to_vec()),
-                    low: 3,
-                    max_size: max_u64,
-                    hit_size_limit: false,
-                    tried_cnt: MAX_ASYNC_FETCH_TRY_CNT,
-                    term: 1,
-                },
-                Ok(4),
-                ents[1..5].to_vec(),
-            ),
-        ];
-
-        for (i, (lo, hi, maxsize, term, async_res, expected_res, expected_ents)) in
-            tests.drain(..).enumerate()
-        {
-            if async_res.low != lo {
-                store.clean_async_fetch_res(lo);
-            } else {
-                store.update_async_fetch_res(lo, Some(Box::new(async_res)));
-            }
-            let mut ents = vec![];
-            store.raft_state.mut_hard_state().set_term(term);
-            let res = store.async_fetch(
-                store.get_region_id(),
-                lo,
-                hi,
-                maxsize,
-                GetEntriesContext::empty(true),
-                &mut ents,
-            );
-            if res != expected_res {
-                panic!("#{}: expect result {:?}, got {:?}", i, expected_res, res);
-            }
-            if ents != expected_ents {
-                panic!("#{}: expect ents {:?}, got {:?}", i, expected_ents, ents);
-            }
-        }
-    }
-
     // last_index and first_index are not mutated by PeerStorage on its own,
     // so we don't test them here.
 
@@ -2765,7 +1659,7 @@ mod tests {
             let (dummy_scheduler, _) = dummy_scheduler();
             let mut store = new_storage_from_ents(sched, dummy_scheduler, &td, &ents);
             let res = store.term(idx).map_err(From::from).and_then(|term| {
-                compact_raft_log(&store.tag, &mut store.entry_storage.apply_state, idx, term)
+                compact_raft_log(&store.tag, store.entry_storage.apply_state_mut(), idx, term)
             });
             // TODO check exact error type after refactoring error.
             if res.is_err() ^ werr.is_err() {
@@ -2886,10 +1780,10 @@ mod tests {
         let mut hs = HardState::default();
         hs.set_commit(7);
         hs.set_term(5);
-        s.raft_state.set_hard_state(hs);
-        s.raft_state.set_last_index(7);
-        s.apply_state.set_applied_index(7);
-        write_task.raft_state = Some(s.raft_state.clone());
+        s.raft_state_mut().set_hard_state(hs);
+        s.raft_state_mut().set_last_index(7);
+        s.apply_state_mut().set_applied_index(7);
+        write_task.raft_state = Some(s.raft_state().clone());
         if write_task.kv_wb.is_none() {
             write_task.kv_wb = Some(s.engines.kv.write_batch());
         }
@@ -2897,7 +1791,7 @@ mod tests {
             .unwrap();
         write_to_db_for_test(&s.engines, write_task);
         let term = s.term(7).unwrap();
-        compact_raft_log(&s.tag, &mut s.entry_storage.apply_state, 7, term).unwrap();
+        compact_raft_log(&s.tag, s.entry_storage.apply_state_mut(), 7, term).unwrap();
         let mut kv_wb = s.engines.kv.write_batch();
         s.save_apply_state_to(&mut kv_wb).unwrap();
         kv_wb.write().unwrap();
@@ -3023,248 +1917,6 @@ mod tests {
     }
 
     #[test]
-    fn test_storage_append() {
-        let ents = vec![new_entry(3, 3), new_entry(4, 4), new_entry(5, 5)];
-        let mut tests = vec![
-            (
-                vec![new_entry(4, 6), new_entry(5, 6)],
-                vec![new_entry(4, 6), new_entry(5, 6)],
-            ),
-            (
-                vec![new_entry(4, 4), new_entry(5, 5), new_entry(6, 5)],
-                vec![new_entry(4, 4), new_entry(5, 5), new_entry(6, 5)],
-            ),
-            // truncate the existing entries and append
-            (vec![new_entry(4, 5)], vec![new_entry(4, 5)]),
-            // direct append
-            (
-                vec![new_entry(6, 5)],
-                vec![new_entry(4, 4), new_entry(5, 5), new_entry(6, 5)],
-            ),
-        ];
-        for (i, (entries, wentries)) in tests.drain(..).enumerate() {
-            let td = Builder::new().prefix("tikv-store-test").tempdir().unwrap();
-            let worker = LazyWorker::new("snap-manager");
-            let sched = worker.scheduler();
-            let (dummy_scheduler, _) = dummy_scheduler();
-            let mut store = new_storage_from_ents(sched, dummy_scheduler, &td, &ents);
-            append_ents(&mut store, &entries);
-            let li = store.last_index().unwrap();
-            let actual_entries = store
-                .entries(4, li + 1, u64::max_value(), GetEntriesContext::empty(false))
-                .unwrap();
-            if actual_entries != wentries {
-                panic!("#{}: want {:?}, got {:?}", i, wentries, actual_entries);
-            }
-        }
-    }
-
-    #[test]
-    fn test_storage_cache_fetch() {
-        let ents = vec![new_entry(3, 3), new_entry(4, 4), new_entry(5, 5)];
-        let td = Builder::new().prefix("tikv-store-test").tempdir().unwrap();
-        let worker = LazyWorker::new("snap-manager");
-        let sched = worker.scheduler();
-        let (dummy_scheduler, _) = dummy_scheduler();
-        let mut store = new_storage_from_ents(sched, dummy_scheduler, &td, &ents);
-        store.cache.cache.clear();
-        // empty cache should fetch data from rocksdb directly.
-        let mut res = store
-            .entries(4, 6, u64::max_value(), GetEntriesContext::empty(false))
-            .unwrap();
-        assert_eq!(*res, ents[1..]);
-
-        let entries = vec![new_entry(6, 5), new_entry(7, 5)];
-        append_ents(&mut store, &entries);
-        validate_cache(&store, &entries);
-
-        // direct cache access
-        res = store
-            .entries(6, 8, u64::max_value(), GetEntriesContext::empty(false))
-            .unwrap();
-        assert_eq!(res, entries);
-
-        // size limit should be supported correctly.
-        res = store
-            .entries(4, 8, 0, GetEntriesContext::empty(false))
-            .unwrap();
-        assert_eq!(res, vec![new_entry(4, 4)]);
-        let mut size: u64 = ents[1..].iter().map(|e| u64::from(e.compute_size())).sum();
-        res = store
-            .entries(4, 8, size, GetEntriesContext::empty(false))
-            .unwrap();
-        let mut exp_res = ents[1..].to_vec();
-        assert_eq!(res, exp_res);
-        for e in &entries {
-            size += u64::from(e.compute_size());
-            exp_res.push(e.clone());
-            res = store
-                .entries(4, 8, size, GetEntriesContext::empty(false))
-                .unwrap();
-            assert_eq!(res, exp_res);
-        }
-
-        // range limit should be supported correctly.
-        for low in 4..9 {
-            for high in low..9 {
-                let res = store
-                    .entries(low, high, u64::max_value(), GetEntriesContext::empty(false))
-                    .unwrap();
-                assert_eq!(*res, exp_res[low as usize - 4..high as usize - 4]);
-            }
-        }
-    }
-
-    #[test]
-    fn test_storage_cache_update() {
-        let ents = vec![new_entry(3, 3), new_entry(4, 4), new_entry(5, 5)];
-        let td = Builder::new().prefix("tikv-store-test").tempdir().unwrap();
-        let worker = LazyWorker::new("snap-manager");
-        let sched = worker.scheduler();
-        let (dummy_scheduler, _) = dummy_scheduler();
-        let mut store = new_storage_from_ents(sched, dummy_scheduler, &td, &ents);
-        store.cache.cache.clear();
-
-        // initial cache
-        let mut entries = vec![new_entry(6, 5), new_entry(7, 5)];
-        append_ents(&mut store, &entries);
-        validate_cache(&store, &entries);
-
-        // rewrite
-        entries = vec![new_entry(6, 6), new_entry(7, 6)];
-        append_ents(&mut store, &entries);
-        validate_cache(&store, &entries);
-
-        // rewrite old entry
-        entries = vec![new_entry(5, 6), new_entry(6, 6)];
-        append_ents(&mut store, &entries);
-        validate_cache(&store, &entries);
-
-        // partial rewrite
-        entries = vec![new_entry(6, 7), new_entry(7, 7)];
-        append_ents(&mut store, &entries);
-        let mut exp_res = vec![new_entry(5, 6), new_entry(6, 7), new_entry(7, 7)];
-        validate_cache(&store, &exp_res);
-
-        // direct append
-        entries = vec![new_entry(8, 7), new_entry(9, 7)];
-        append_ents(&mut store, &entries);
-        exp_res.extend_from_slice(&entries);
-        validate_cache(&store, &exp_res);
-
-        // rewrite middle
-        entries = vec![new_entry(7, 8)];
-        append_ents(&mut store, &entries);
-        exp_res.truncate(2);
-        exp_res.push(new_entry(7, 8));
-        validate_cache(&store, &exp_res);
-
-        // compact to min(5 + 1, 7)
-        store.cache.persisted = 5;
-        store.compact_entry_cache(7);
-        exp_res = vec![new_entry(6, 7), new_entry(7, 8)];
-        validate_cache(&store, &exp_res);
-
-        // compact to min(7 + 1, 7)
-        store.cache.persisted = 7;
-        store.compact_entry_cache(7);
-        exp_res = vec![new_entry(7, 8)];
-        validate_cache(&store, &exp_res);
-        // compact all
-        store.compact_entry_cache(8);
-        validate_cache(&store, &[]);
-        // invalid compaction should be ignored.
-        store.compact_entry_cache(6);
-    }
-
-    #[test]
-    fn test_storage_cache_size_change() {
-        let new_padded_entry = |index: u64, term: u64, pad_len: usize| {
-            let mut e = new_entry(index, term);
-            e.data = vec![b'x'; pad_len].into();
-            e
-        };
-
-        // Test the initial data structure size.
-        let (tx, rx) = mpsc::sync_channel(8);
-        let mut cache = EntryCache::new_with_cb(move |c: i64| tx.send(c).unwrap());
-        assert_eq!(rx.try_recv().unwrap(), 896);
-
-        cache.append(
-            0,
-            0,
-            &[new_padded_entry(101, 1, 1), new_padded_entry(102, 1, 2)],
-        );
-        assert_eq!(rx.try_recv().unwrap(), 3);
-
-        // Test size change for one overlapped entry.
-        cache.append(0, 0, &[new_padded_entry(102, 2, 3)]);
-        assert_eq!(rx.try_recv().unwrap(), 1);
-
-        // Test size change for all overlapped entries.
-        cache.append(
-            0,
-            0,
-            &[new_padded_entry(101, 3, 4), new_padded_entry(102, 3, 5)],
-        );
-        assert_eq!(rx.try_recv().unwrap(), 5);
-
-        cache.append(0, 0, &[new_padded_entry(103, 3, 6)]);
-        assert_eq!(rx.try_recv().unwrap(), 6);
-
-        // Test trace a dangle entry.
-        let cached_entries = CachedEntries::new(vec![new_padded_entry(100, 1, 1)]);
-        cache.trace_cached_entries(cached_entries);
-        assert_eq!(rx.try_recv().unwrap(), 1);
-
-        // Test trace an entry which is still in cache.
-        let cached_entries = CachedEntries::new(vec![new_padded_entry(102, 3, 5)]);
-        cache.trace_cached_entries(cached_entries);
-        assert_eq!(rx.try_recv().unwrap(), 0);
-
-        // Test compare `cached_last` with `trunc_to_idx` in `EntryCache::append_impl`.
-        cache.append(0, 0, &[new_padded_entry(103, 4, 7)]);
-        assert_eq!(rx.try_recv().unwrap(), 1);
-
-        // Test compact one traced dangle entry and one entry in cache.
-        cache.persisted = 101;
-        cache.compact_to(102);
-        assert_eq!(rx.try_recv().unwrap(), -5);
-
-        // Test compact the last traced dangle entry.
-        cache.persisted = 102;
-        cache.compact_to(103);
-        assert_eq!(rx.try_recv().unwrap(), -5);
-
-        // Test compact all entries.
-        cache.persisted = 103;
-        cache.compact_to(104);
-        assert_eq!(rx.try_recv().unwrap(), -7);
-
-        drop(cache);
-        assert_eq!(rx.try_recv().unwrap(), -896);
-    }
-
-    #[test]
-    fn test_storage_cache_entry() {
-        let mut cache = EntryCache::default();
-        let ents = vec![
-            new_entry(3, 3),
-            new_entry(4, 4),
-            new_entry(5, 4),
-            new_entry(6, 6),
-        ];
-        cache.append(0, 0, &ents);
-        assert!(cache.entry(1).is_none());
-        assert!(cache.entry(2).is_none());
-        for e in &ents {
-            assert_eq!(e, cache.entry(e.get_index()).unwrap());
-        }
-        let res = panic_hook::recover_safe(|| cache.entry(7));
-        assert!(res.is_err());
-    }
-
-    #[test]
     fn test_storage_apply_snapshot() {
         let ents = vec![
             new_entry(3, 3),
@@ -3316,11 +1968,11 @@ mod tests {
         let mut snap_data = RaftSnapshotData::default();
         snap_data.merge_from_bytes(snap1.get_data()).unwrap();
         assert_eq!(snap_region, snap_data.take_region(),);
-        assert_eq!(s2.last_term, snap1.get_metadata().get_term());
-        assert_eq!(s2.apply_state.get_applied_index(), 6);
-        assert_eq!(s2.raft_state.get_last_index(), 6);
-        assert_eq!(s2.apply_state.get_truncated_state().get_index(), 6);
-        assert_eq!(s2.apply_state.get_truncated_state().get_term(), 6);
+        assert_eq!(s2.last_term(), snap1.get_metadata().get_term());
+        assert_eq!(s2.apply_state().get_applied_index(), 6);
+        assert_eq!(s2.raft_state().get_last_index(), 6);
+        assert_eq!(s2.apply_state().get_truncated_state().get_index(), 6);
+        assert_eq!(s2.apply_state().get_truncated_state().get_term(), 6);
         assert_eq!(s2.first_index(), Ok(s2.applied_index() + 1));
         validate_cache(&s2, &[]);
 
@@ -3333,11 +1985,11 @@ mod tests {
         let mut snap_data = RaftSnapshotData::default();
         snap_data.merge_from_bytes(snap1.get_data()).unwrap();
         assert_eq!(snap_region, snap_data.take_region(),);
-        assert_eq!(s3.last_term, snap1.get_metadata().get_term());
-        assert_eq!(s3.apply_state.get_applied_index(), 6);
-        assert_eq!(s3.raft_state.get_last_index(), 6);
-        assert_eq!(s3.apply_state.get_truncated_state().get_index(), 6);
-        assert_eq!(s3.apply_state.get_truncated_state().get_term(), 6);
+        assert_eq!(s3.last_term(), snap1.get_metadata().get_term());
+        assert_eq!(s3.apply_state().get_applied_index(), 6);
+        assert_eq!(s3.raft_state().get_last_index(), 6);
+        assert_eq!(s3.apply_state().get_truncated_state().get_index(), 6);
+        assert_eq!(s3.apply_state().get_truncated_state().get_term(), 6);
         validate_cache(&s3, &[]);
     }
